efiboot: add file existance checking
Currently the efiboot file is not checked for existance, which might
lead to boot failure if continue to install while file not present.

+++ sbin/functions.sh
@@ -622,7 +622,7 @@ install_grub()
 	
 	#install efi boot
 
-	if [ -n "${INSTALL_EFIBOOT}" ]; then
+	if [ -n "${INSTALL_EFIBOOT}" ] && [ -e "${INSTALL_EFIBOOT}" ]; then
 		debugmsg ${DEBUG_INFO} "Installing the EFI bootloader"
 		mkdir -p ${mountpoint}/EFI/BOOT/
 		cp $INSTALL_EFIBOOT ${mountpoint}/EFI/BOOT/
